requirements.txt
# FastAPI Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Supabase
supabase==2.3.4
postgrest==0.13.2

# Authentication & Security
pyjwt==2.8.0
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
cryptography==42.0.0

# Google AI
google-generativeai==0.3.2

# HTTP & Requests
httpx==0.26.0
requests==2.31.0
aiohttp==3.9.1

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Environment Variables
python-dotenv==1.0.0

# Date & Time
python-dateutil==2.8.2

# Rate Limiting
slowapi==0.1.9

# CORS
python-cors==1.0.0

# Redis (optionnel pour cache)
redis==5.0.1

# Monitoring & Logs
loguru==0.7.2

# SQLAlchemy for local PostgreSQL
sqlalchemy==2.0.23
psycopg2-binary==2.9.9

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
